Refactor handlers into seperate files
