+++ src/handlers/index.ts
@@ -1,0 +1,2 @@
+export { getCountries } from "./country.handler";
+export { getEmissions } from "./emission.handler";
